# openapi.yaml
openapi: 3.0.3
info:
  title: Portal de Benefícios KNN API
  description: |
    API para o Portal de Benefícios KNN, permitindo que alunos da KNN Idiomas acessem promoções
    exclusivas de parceiros comerciais através de códigos de validação temporários.
  version: 1.0.0
  contact:
    name: Equipe de TI KNN Idiomas
    email: ti@knnidiomas.com.br
servers:
  - url: https://api.knnidiomas.com.br/v1
    description: Ambiente de produção
  - url: http://localhost:8000/v1
    description: Ambiente de desenvolvimento local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT com claim 'role' (student|partner|admin)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CODE
            msg:
              type: string
              example: Código expirado ou já usado
          required:
            - code
            - msg
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
          description: Dados da resposta (varia conforme endpoint)
        msg:
          type: string
          default: ok
      required:
        - data
        - msg
...
    Promotion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        title:
          type: string
          example: 20% de desconto em livros de idiomas
        type:
          type: string
          example: discount
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
      required:
        - id
        - partner_id
        - title
        - type
        - valid_from
        - valid_to
        - active
...
    RedeemRequest:
      type: object
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
          example: '123456'
        cpf:
          type: string
          pattern: '^[0-9]{11}$'
          example: '12345678901'
      required:
        - code
        - cpf
...
    PromotionRequest:
      type: object
      properties:
        title:
          type: string
          example: 20% de desconto em livros de idiomas
        type:
          type: string
          example: discount
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
      required:
        - title
        - type
        - valid_from
        - valid_to
...
  parameters:
...
    PaginationLimit:
      name: limit
      in: query
      description: Número máximo de itens por página
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    PaginationOffset:
      name: offset
      in: query
      description: Offset para paginação
      schema:
        type: integer
        default: 0
        minimum: 0
    DateRangeParam:
      name: range
      in: query
      description: Período para relatório (formato YYYY-MM)
      required: true
      schema:
        type: string
        pattern: '^[0-9]{4}-[0-9]{2}$'
        example: '2025-05'
...
    IdParam:
      name: id
      in: path
      description: ID do recurso
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - bearerAuth: []
paths:
...
  /partner/redeem:
    post:
      summary: Resgata código de validação
      description: Resgata um código de validação gerado por um aluno
      tags:
        - partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemRequest'
      responses:
        '200':
          description: Código resgatado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          student:
                            type: object
                            properties:
                              name:
                                type: string
                              course:
                                type: string
                            required:
                              - name
                              - course
                          promotion:
                            type: object
                            properties:
                              title:
                                type: string
                            required:
                              - title
                        required:
                          - student
                          - promotion
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Código não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Código expirado ou promoção inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /partner/promotions:
    get:
      summary: Lista promoções do parceiro
      description: Retorna a lista de promoções do parceiro
      tags:
        - partner
      parameters:
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationOffset'
      responses:
        '200':
          description: Lista de promoções
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Promotion'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Cria nova promoção
      description: Cria uma nova promoção para o parceiro
      tags:
        - partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
      responses:
        '201':
          description: Promoção criada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Promotion'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /partner/promotions/{id}:
    put:
      summary: Atualiza promoção
      description: Atualiza uma promoção existente do parceiro
      tags:
        - partner
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
      responses:
        '200':
          description: Promoção atualizada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Promotion'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Promoção não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Desativa promoção
      description: Desativa uma promoção existente do parceiro
      tags:
        - partner
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Promoção desativada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Promoção não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /partner/reports:
    get:
      summary: Relatório de uso
      description: Retorna relatório de uso das promoções do parceiro
      tags:
        - partner
      parameters:
        - $ref: '#/components/parameters/DateRangeParam'
      responses:
        '200':
          description: Relatório de uso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          period:
                            type: string
                            example: '2025-05'
                          total_codes:
                            type: integer
                          total_redemptions:
                            type: integer
                          promotions:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                title:
                                  type: string
                                redemptions:
                                  type: integer
                              required:
                                - id
                                - title
                                - redemptions
                        required:
                          - period
                          - total_codes
                          - total_redemptions
                          - promotions
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Acesso negado (role inválida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Parâmetro de data inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...